Crate seccompiler_bin

seccompiler-bin is a program that compiles multi-threaded seccomp-bpf filters expressed as JSON into raw BPF programs, serializing them and outputting them to a file.

Used in conjunction with the provided library crate, one can deserialize the binary filters and easily install them on a per-thread basis, in order to achieve a quick and robust seccomp-based jailing solution.

See the documentation on github for more information.

The compilation goes through a couple of steps, from JSON to BPF:

                JSON
                 |
          (via serde_json)
                 |
                 V
     collection of `Filter` objects
                 |
    (via Compiler.compile_blob(...))
                 |
                 V
 collection of `SeccompFilter` objects
   (IR - intermediate representation)
                 |
  (via SeccompFilter.try_into::<BpfProgram>(...))
                 |
                 V
   collection of `BpfProgram` objects

Modules

backend 🔒

This module defines the data structures used for the intermmediate representation (IR), as well as the logic for compiling the filter into BPF code, the final form of the filter.

common 🔒

Module that defines common data structures used by both the library crate and seccompiler-bin.

compiler 🔒

Module defining the logic for compiling the deserialized filter objects into the IR. Used by seccompiler-bin.

syscall_table 🔒

Structs

Arguments 🔒

Enums

SeccompError 🔒

SeccompilerError 🔒

Constants

DEFAULT_OUTPUT_FILENAME 🔒

SECCOMPILER_VERSION 🔒

_DERIVE_Display_FOR_SeccompilerError 🔒

Functions

build_arg_parser 🔒

compile 🔒

get_argument_values 🔒

main 🔒

main_exec 🔒